+++ src/App.css
@@ -6,9 +6,3 @@
 html {
   scroll-behavior: smooth;
 }
-/* body {
-  @media screen and (min-width: 600px) {
-        font-size: 14px;
-    }
-} */
-
