+++ src/components/Theme.js
@@ -1,61 +1,3 @@
-import React, { useState } from "react";
-import { ThemeProvider } from "styled-components";
-
-
-
-// const theme = {
-//     colors: {
-//         dephtsOfWater: {
-//             stPatrickBlue: 'rgb(22, 31, 111)',
-//             stPatrickTintOne: '#A2A5C5',
-//             stPatrickTintTwo:'#B9BCD4',
-//             stPatrickTintThree:'#D0D2E2',
-//             stPatrickTintFour:'#E8E9F1',
-//             denimBlue: 'rgb(33, 64, 163)',
-//             hanBLue: 'rgb(54, 105, 213)',
-//             bleuDeFrance: 'rgb(36, 136, 213)',
-//             pacificBlue: 'rgb(26, 169, 208)',
-//         }, 
-//         blueberry: {
-//             spaceCadet: 'rgb(31, 48, 80)',
-//             americanBlue: 'rgb(42, 65, 111)',
-//             queenBlue: 'rgb(73, 104, 152)',
-//             darkPastelBlue: 'rgb(117, 152, 200)',
-//         },
-//         darkOrange: {
-//             secondary: 'rgb(1, 0, 19)',
-//             secondaryAlt: '#1A1A2B',
-//             secondaryAlt2: '#0E0D1E',
-//             tertiaryAlt: 'rgb(18, 21, 76)',
-//             stPatrickBlue: 'rgb(22, 31, 111)',
-//             tertiary: 'rgb(24, 135, 199)',
-//             primaryAlt: 'rgb(255, 119, 4)',
-//             primary: 'rgb(246, 83, 15)',
-//         }
-        // darkOrange: {
-        //     richBlack: 'rgb(1, 0, 19)',
-        //     richBlackTint: '#1A1A2B',
-        //     richBlackTone: '#0E0D1E',
-        //     cetaceanBlue: 'rgb(18, 21, 76)',
-        //     stPatrickBlue: 'rgb(22, 31, 111)',
-        //     cyanCornflowerBlue: 'rgb(24, 135, 199)',
-        //     heatWave: 'rgb(255, 119, 4)',
-        //     oriolesOrange: 'rgb(246, 83, 15)',
-        // }
-//     }
-// }
-
-// const Theme = ({ themeColors, setThemeColors, children }) => {
-
-//     setThemeColors(theme.colors.darkOrange)
-
-//     return (
-//         <ThemeProvider theme={themeColors}>{children}</ThemeProvider>
-//     );
-// } 
-
-// export default Theme;
-
 
 
 export const thin = {
