+++ src/components/index.js
@@ -1,0 +1,12 @@
+export { default as Navbar } from './navbar'
+export { default as Footer } from './footer'
+
+export { default as CardList } from './card-list'
+export { default as Alert } from './alert'
+export { default as Loading } from './loading'
+export { default as SearchCollectionCard } from './search-collection-card'
+export { default as SearchBar } from './searchbar'
+export { default as NewsCard } from './news-card'
+export { default as DiscoverCategory } from './discover-category'
+
+export { default as ToTopButton } from './to-top-button'
