--- src/constants/colors.js
@@ -1,9 +1,0 @@
-// let russianVioletDark = '#10002B'
-// let russianVioletLightDark = '#240046'
-// let russianShade = '#CE99FF'
-// let persionIndigo = '#3C096C'
-// let purple = '#5A189A'
-// let frenchViolet = '#7B2CBF'
-// let darkOrchid = '#9D4EDD'
-// let heliotrope = '#C77DFF'
-// let mauve = '#E0AAFF'
