--- src/hooks/use-storage-listener.js
@@ -1,15 +1,0 @@
-import { useState } from 'react'
-
-
-// export default function useStorageListener() {
-
-//     const feed = 'feed'
-//     const [feedStorage, setFeedStorage] = useState(JSON.parse(localStorage.getItem(feed)))
-
-//     const updateFeeds = (feeds) => {
-        
-//     }
-
-
-//     return { feedStorage }
-// }
