+++ src/pages/home.js
@@ -1,17 +1,13 @@
 import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
 import { getParsedResponse } from '../services/parser'
-import { useStateWithPromise } from '../helpers/use-state-with-promise'
+import { useStateWithPromise } from '../hooks/use-state-with-promise'
 
-import { CardList } from '../components/card-list'
-import { Alert } from '../components/alert'
-import { Loading } from '../components/loading'
-import SearchCollectionCard from '../components/search-collection-card'
+import { CardList, Alert, Loading, SearchCollectionCard, SearchBar } from '../components'
 
 import DeleteForeverIcon from '@material-ui/icons/DeleteForever';
-import SearchBar from '../components/searchbar'
 
-export const Home = () => {
+export default function Home() {
 
     const feedStr = 'rss-feed-manager'
     const [feeds, setFeeds] = useState(JSON.parse(localStorage.getItem(feedStr)))
