+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export { default as Home} from './home';
+export { default as SearchFeed} from './search-feed';
+export { default as DiscoverFeed} from './discover-feed';
+export { default as NotFound} from './not-found';
